Rename zookeeper-curator to zookeeper (#21936)
* Rename zookeeper-curator to zookeeper

* Rename zookeeper-curator to zookeeper

* Rename zookeeper-curator to zookeeper
