+++ examples/shardingsphere-jdbc-example/single-feature-example/cluster-mode-example/cluster-mode-raw-jdbc-example/src/main/java/org/apache/shardingsphere/example/cluster/mode/raw/jdbc/ClusterModeRawJavaConfigurationExample.java
@@ -40,7 +40,7 @@ import java.sql.SQLException;
 
 /*
  * 1. Please make sure primary replica data replication sync on MySQL is running correctly. Otherwise this example will query empty data from replica.
- * 2. Please make sure sharding-governance-center-zookeeper-curator in your pom if registryCenterType = RegistryCenterType.ZOOKEEPER.
+ * 2. Please make sure sharding-governance-center-zookeeper in your pom if registryCenterType = RegistryCenterType.ZOOKEEPER.
  */
 public final class ClusterModeRawJavaConfigurationExample {
