--- mode/type/cluster/repository/provider/zookeeper-curator/src/test/java/org/apache/shardingsphere/mode/repository/cluster/zookeeper/handler/CuratorZookeeperExceptionHandlerTest.java
@@ -1,44 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.mode.repository.cluster.zookeeper.handler;
-
-import org.apache.shardingsphere.mode.repository.cluster.exception.ClusterPersistRepositoryException;
-import org.apache.zookeeper.KeeperException.NoNodeException;
-import org.junit.Test;
-
-import static org.hamcrest.CoreMatchers.instanceOf;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.Assert.fail;
-
-public final class CuratorZookeeperExceptionHandlerTest {
-    
-    @Test
-    public void assertHandleException() {
-        CuratorZookeeperExceptionHandler.handleException(null);
-        CuratorZookeeperExceptionHandler.handleException(new NoNodeException(""));
-        CuratorZookeeperExceptionHandler.handleException(new Exception(new NoNodeException("")));
-        try {
-            CuratorZookeeperExceptionHandler.handleException(new Exception(""));
-            fail("must be failed handle new Exception().");
-            // CHECKSTYLE:OFF
-        } catch (final Exception ex) {
-            // CHECKSTYLE:ON
-            assertThat(ex, instanceOf(ClusterPersistRepositoryException.class));
-        }
-    }
-}
