+++ mode/type/cluster/repository/provider/zookeeper/pom.xml
@@ -21,17 +21,35 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.apache.shardingsphere</groupId>
-        <artifactId>shardingsphere-cluster-mode-repository</artifactId>
+        <artifactId>shardingsphere-cluster-mode-repository-provider</artifactId>
         <version>5.2.2-SNAPSHOT</version>
     </parent>
-    <artifactId>shardingsphere-cluster-mode-repository-provider</artifactId>
-    <packaging>pom</packaging>
+    <artifactId>shardingsphere-cluster-mode-repository-zookeeper</artifactId>
     <name>${project.artifactId}</name>
     
-    <modules>
-        <module>zookeeper</module>
-        <module>etcd</module>
-        <module>nacos</module>
-        <module>consul</module>
-    </modules>
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.shardingsphere</groupId>
+            <artifactId>shardingsphere-cluster-mode-repository-api</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        
+        <dependency>
+            <groupId>org.apache.curator</groupId>
+            <artifactId>curator-framework</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.curator</groupId>
+            <artifactId>curator-client</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.curator</groupId>
+            <artifactId>curator-recipes</artifactId>
+        </dependency>
+        
+        <dependency>
+            <groupId>org.apache.curator</groupId>
+            <artifactId>curator-test</artifactId>
+        </dependency>
+    </dependencies>
 </project>
