+++ test/integration-test/scaling/src/test/java/org/apache/shardingsphere/integration/data/pipeline/framework/watcher/ScalingWatcher.java
@@ -26,7 +26,7 @@ import org.apache.shardingsphere.integration.data.pipeline.framework.container.c
 import org.apache.shardingsphere.integration.data.pipeline.framework.container.compose.NativeContainerComposer;
 import org.apache.shardingsphere.mode.repository.cluster.ClusterPersistRepository;
 import org.apache.shardingsphere.mode.repository.cluster.ClusterPersistRepositoryConfiguration;
-import org.apache.shardingsphere.mode.repository.cluster.zookeeper.CuratorZookeeperRepository;
+import org.apache.shardingsphere.mode.repository.cluster.zookeeper.ZookeeperRepository;
 import org.junit.rules.TestWatcher;
 import org.junit.runner.Description;
 
@@ -57,7 +57,7 @@ public class ScalingWatcher extends TestWatcher {
         String namespace = "it_db_" + databaseType.getType().toLowerCase();
         ClusterPersistRepositoryConfiguration config = new ClusterPersistRepositoryConfiguration("ZooKeeper", namespace,
                 dockerContainerComposer.getGovernanceContainer().getServerLists(), new Properties());
-        ClusterPersistRepository zookeeperRepository = new CuratorZookeeperRepository();
+        ClusterPersistRepository zookeeperRepository = new ZookeeperRepository();
         zookeeperRepository.init(config, new ProxyInstanceMetaData(UUID.randomUUID().toString(), 3307));
         List<String> childrenKeys = zookeeperRepository.getChildrenKeys("/");
         for (String each : childrenKeys) {
